fix(wac): usar DECIMAL(16,4) en wac_huevos_inventario para consistencia con legacy
La columna original se definió como BIGINT UNSIGNED asumiendo que siempre
almacenaríamos enteros. Al implementar el listener dual-write de Fase 2 se
descubrió que el flujo real acumula valores fraccionarios: la columna legacy
huevos_facturados_acumulados es DECIMAL(14,2) y viaje_cargas.cantidad es
DECIMAL(14,3), y la conversión cartones * huevos_por_carton puede producir
fracciones (ej: 10.55 * 30 = 316.5).

Mantener BIGINT UNSIGNED habría forzado un truncamiento sistemático que
divergiría del legacy y rompería la paridad que el shadow mode busca
validar. DECIMAL(16,4) da headroom suficiente (10^12 huevos) con 4
decimales de precisión.

La migración es idempotente (detecta columnas existentes) por el patrón
ya establecido en Fase 0.

+++ database/migrations/2026_04_22_000001_add_wac_columns_to_lotes_table.php
@@ -43,9 +43,13 @@ return new class extends Migration {
                 ->comment('WAC: valor monetario del stock facturado actual (numerador del costo promedio perpetuo)');
 
             // Denominador WAC: cantidad de huevos facturados actualmente en stock
-            // BIGINT UNSIGNED → headroom futuro si el volumen del negocio crece 10x-100x
-            // Se guarda solo huevos facturados (no incluye regalo, que no tiene costo)
-            $table->unsignedBigInteger('wac_huevos_inventario')
+            // DECIMAL(16,4) → consistente con wac_costo_inventario y con la columna legacy
+            // huevos_facturados_acumulados DECIMAL(14,2). Se necesitan decimales porque cartones
+            // es DECIMAL(14,2) y viaje_cargas.cantidad es DECIMAL(14,3); la conversión
+            // cartones × huevos_por_carton puede producir fracciones (ej: 10.55 × 30 = 316.5).
+            // Usar entero truncaría precisión y generaría divergencia sistemática contra legacy.
+            // Se guarda solo huevos facturados (no incluye regalo, que no tiene costo).
+            $table->decimal('wac_huevos_inventario', 16, 4)
                 ->nullable()
                 ->after('wac_costo_inventario')
                 ->comment('WAC: cantidad de huevos facturados actualmente en stock (denominador del costo promedio perpetuo)');
